+++ routes/adminbackend.php
@@ -10,4 +10,9 @@ Route::get('/backend/admin/getProfile', [ProfileController::class, 'backend_getA
 
 // Admin Management
 Route::get('/backend/admin/getAdmin', [AdminManagementController::class, 'backend_getAdmin']);
+
+Route::post('/backend/admin/addOffice', [AdminManagementController::class, 'backend_addOffice']);
 Route::get('/backend/admin/getOffice', [AdminManagementController::class, 'backend_getOffice']);
+Route::get('/backend/admin/getOneOffice/{office_id}', [AdminManagementController::class, 'backend_getOneOffice']);
+Route::put('/backend/admin/editOffice/{office_id}', [AdminManagementController::class, 'backend_editOffice']);
+Route::delete('/backend/admin/removeOffice/{office_id}', [AdminManagementController::class, 'backend_removeOffice']);